# backend/tests/contract/openapi.json
{
  "components": {
    "schemas": {
      "And": {
        "properties": {
          "children": {
            "items": {
              "discriminator": {
                "mapping": {
                  "and": "#/components/schemas/And",
                  "filter": "#/components/schemas/Filter",
                  "near": "#/components/schemas/Near",
                  "not": "#/components/schemas/Not",
                  "or": "#/components/schemas/Or",
                  "phrase": "#/components/schemas/Phrase",
                  "term": "#/components/schemas/Term",
                  "wildcard": "#/components/schemas/Wildcard"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Term"
                },
                {
                  "$ref": "#/components/schemas/Wildcard"
                },
                {
                  "$ref": "#/components/schemas/Phrase"
                },
                {
                  "$ref": "#/components/schemas/Near"
                },
                {
                  "$ref": "#/components/schemas/Filter"
                },
                {
                  "$ref": "#/components/schemas/Not"
                },
                {
                  "$ref": "#/components/schemas/And"
                },
                {
                  "$ref": "#/components/schemas/Or"
                }
              ]
            },
            "minItems": 2,
            "title": "Children",
            "type": "array"
          },
          "kind": {
            "const": "and",
            "default": "and",
            "title": "Kind",
            "type": "string"
          },
          "span": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "integer"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Span",
            "type": "array"
          }
        },
        "required": [
          "span",
          "kind",
          "children"
        ],
        "title": "And",
        "type": "object"
      },
...
      "Diagnostic": {
        "description": "A warning, error or translation notice about a query, with a half-open code-point span into `q`.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/DiagnosticCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "span": {
            "anyOf": [
              {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "type": "integer"
                  },
                  {
                    "type": "integer"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Span"
          }
        },
        "required": [
          "code",
          "message",
          "span"
        ],
        "title": "Diagnostic",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "Filter": {
        "description": "`field:` restricted to any of `values` (a single-field OR group): YearRanges for `year`, canonical\nstrings (`vocab.py`) for the others.",
        "properties": {
          "field": {
            "description": "Open set: new values may be added within /api/v1; handle a value you don't know.",
            "enum": [
              "venue",
              "year",
              "track",
              "status"
            ],
            "title": "Field",
            "type": "string"
          },
          "kind": {
            "const": "filter",
            "default": "filter",
            "title": "Kind",
            "type": "string"
          },
          "span": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "integer"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Span",
            "type": "array"
          },
          "values": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "$ref": "#/components/schemas/YearRange"
                }
              ]
            },
            "minItems": 1,
            "title": "Values",
            "type": "array"
          }
        },
        "required": [
          "span",
          "kind",
          "field",
          "values"
        ],
        "title": "Filter",
        "type": "object"
      },
...
      "Near": {
        "description": "Both operands within `distance` words of each other, in either order, in one field.",
        "properties": {
          "distance": {
            "maximum": 100,
            "minimum": 0,
            "title": "Distance",
            "type": "integer"
          },
          "kind": {
            "const": "near",
            "default": "near",
            "title": "Kind",
            "type": "string"
          },
          "left": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Term"
              },
              {
                "$ref": "#/components/schemas/Wildcard"
              },
              {
                "$ref": "#/components/schemas/Phrase"
              }
            ],
            "title": "Left"
          },
          "right": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Term"
              },
              {
                "$ref": "#/components/schemas/Wildcard"
              },
              {
                "$ref": "#/components/schemas/Phrase"
              }
            ],
            "title": "Right"
          },
          "span": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "integer"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Span",
            "type": "array"
          }
        },
        "required": [
          "span",
          "kind",
          "left",
          "right",
          "distance"
        ],
        "title": "Near",
        "type": "object"
      },
      "Not": {
        "properties": {
          "child": {
            "discriminator": {
              "mapping": {
                "and": "#/components/schemas/And",
                "filter": "#/components/schemas/Filter",
                "near": "#/components/schemas/Near",
                "not": "#/components/schemas/Not",
                "or": "#/components/schemas/Or",
                "phrase": "#/components/schemas/Phrase",
                "term": "#/components/schemas/Term",
                "wildcard": "#/components/schemas/Wildcard"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/Term"
              },
              {
                "$ref": "#/components/schemas/Wildcard"
              },
              {
                "$ref": "#/components/schemas/Phrase"
              },
              {
                "$ref": "#/components/schemas/Near"
              },
              {
                "$ref": "#/components/schemas/Filter"
              },
              {
                "$ref": "#/components/schemas/Not"
              },
              {
                "$ref": "#/components/schemas/And"
              },
              {
                "$ref": "#/components/schemas/Or"
              }
            ],
            "title": "Child"
          },
          "kind": {
            "const": "not",
            "default": "not",
            "title": "Kind",
            "type": "string"
          },
          "span": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "integer"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Span",
            "type": "array"
          }
        },
        "required": [
          "span",
          "kind",
          "child"
        ],
        "title": "Not",
        "type": "object"
      },
      "Or": {
        "properties": {
          "children": {
            "items": {
              "discriminator": {
                "mapping": {
                  "and": "#/components/schemas/And",
                  "filter": "#/components/schemas/Filter",
                  "near": "#/components/schemas/Near",
                  "not": "#/components/schemas/Not",
                  "or": "#/components/schemas/Or",
                  "phrase": "#/components/schemas/Phrase",
                  "term": "#/components/schemas/Term",
                  "wildcard": "#/components/schemas/Wildcard"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Term"
                },
                {
                  "$ref": "#/components/schemas/Wildcard"
                },
                {
                  "$ref": "#/components/schemas/Phrase"
                },
                {
                  "$ref": "#/components/schemas/Near"
                },
                {
                  "$ref": "#/components/schemas/Filter"
                },
                {
                  "$ref": "#/components/schemas/Not"
                },
                {
                  "$ref": "#/components/schemas/And"
                },
                {
                  "$ref": "#/components/schemas/Or"
                }
              ]
            },
            "minItems": 2,
            "title": "Children",
            "type": "array"
          },
          "kind": {
            "const": "or",
            "default": "or",
            "title": "Kind",
            "type": "string"
          },
          "span": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "integer"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Span",
            "type": "array"
          }
        },
        "required": [
          "span",
          "kind",
          "children"
        ],
        "title": "Or",
        "type": "object"
      },
...
      "ParseRequest": {
        "additionalProperties": false,
        "properties": {
          "mode": {
            "default": "native",
            "description": "`native` (this grammar) or `scholar` (Google Scholar / Publish or Perish syntax, translated).",
            "enum": [
              "native",
              "scholar"
            ],
            "title": "Mode",
            "type": "string"
          },
          "q": {
            "description": "The query (spec 02 grammar). At most 2,000 Unicode code points: a longer one is 422 `PARSE_TOO_LONG`, refused before it is parsed; so is one whose canonical form (defaults written out) is longer, refused after canonicalising.",
            "title": "Q",
            "type": "string"
          }
        },
        "required": [
          "q"
        ],
        "title": "ParseRequest",
        "type": "object"
      },
      "ParseResponse": {
        "description": "02's `ParseResult` without `identification_ast` (it stays server-side; spec 04 §Endpoints). A query\nwith errors is still a 200 here: `errors` holds them, and every Optional is null.",
        "properties": {
          "ast": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "and": "#/components/schemas/And",
                    "filter": "#/components/schemas/Filter",
                    "near": "#/components/schemas/Near",
                    "not": "#/components/schemas/Not",
                    "or": "#/components/schemas/Or",
                    "phrase": "#/components/schemas/Phrase",
                    "term": "#/components/schemas/Term",
                    "wildcard": "#/components/schemas/Wildcard"
                  },
                  "propertyName": "kind"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/Term"
                  },
                  {
                    "$ref": "#/components/schemas/Wildcard"
                  },
                  {
                    "$ref": "#/components/schemas/Phrase"
                  },
                  {
                    "$ref": "#/components/schemas/Near"
                  },
                  {
                    "$ref": "#/components/schemas/Filter"
                  },
                  {
                    "$ref": "#/components/schemas/Not"
                  },
                  {
                    "$ref": "#/components/schemas/And"
                  },
                  {
                    "$ref": "#/components/schemas/Or"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Ast"
          },
          "canonical": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical"
          },
          "canonical_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Hash"
          },
          "defaults": {
            "items": {
              "description": "Open set: new values may be added within /api/v1; handle a value you don't know.",
              "enum": [
                "venue",
                "year",
                "track",
                "status"
              ],
              "type": "string"
            },
            "title": "Defaults",
            "type": "array"
          },
          "effective_ast": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "and": "#/components/schemas/And",
                    "filter": "#/components/schemas/Filter",
                    "near": "#/components/schemas/Near",
                    "not": "#/components/schemas/Not",
                    "or": "#/components/schemas/Or",
                    "phrase": "#/components/schemas/Phrase",
                    "term": "#/components/schemas/Term",
                    "wildcard": "#/components/schemas/Wildcard"
                  },
                  "propertyName": "kind"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/Term"
                  },
                  {
                    "$ref": "#/components/schemas/Wildcard"
                  },
                  {
                    "$ref": "#/components/schemas/Phrase"
                  },
                  {
                    "$ref": "#/components/schemas/Near"
                  },
                  {
                    "$ref": "#/components/schemas/Filter"
                  },
                  {
                    "$ref": "#/components/schemas/Not"
                  },
                  {
                    "$ref": "#/components/schemas/And"
                  },
                  {
                    "$ref": "#/components/schemas/Or"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Ast"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            },
            "title": "Errors",
            "type": "array"
          },
          "identification_query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identification Query"
          },
          "index_version": {
            "title": "Index Version",
            "type": "string"
          },
          "mode": {
            "enum": [
              "native",
              "scholar"
            ],
            "title": "Mode",
            "type": "string"
          },
          "query_version": {
            "title": "Query Version",
            "type": "string"
          },
          "tokenizer_version": {
            "title": "Tokenizer Version",
            "type": "string"
          },
          "translations": {
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            },
            "title": "Translations",
            "type": "array"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/Diagnostic"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "index_version",
          "tokenizer_version",
          "query_version",
          "mode",
          "ast",
          "effective_ast",
          "canonical",
          "canonical_hash",
          "identification_query",
          "defaults",
          "warnings",
          "errors",
          "translations"
        ],
        "title": "ParseResponse",
        "type": "object"
      },
      "Phrase": {
        "description": "Consecutive positions in one field. A wildcard item is expanded at its position (decision-001).",
        "properties": {
          "field": {
            "anyOf": [
              {
                "description": "Open set: new values may be added within /api/v1; handle a value you don't know.",
                "enum": [
                  "title",
                  "abstract"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "items": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Term"
                },
                {
                  "$ref": "#/components/schemas/Wildcard"
                }
              ]
            },
            "minItems": 2,
            "title": "Items",
            "type": "array"
          },
          "kind": {
            "const": "phrase",
            "default": "phrase",
            "title": "Kind",
            "type": "string"
          },
          "span": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "integer"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Span",
            "type": "array"
          }
        },
        "required": [
          "span",
          "kind",
          "items",
          "field"
        ],
        "title": "Phrase",
        "type": "object"
      },
...
      "SearchResponse": {
        "properties": {
          "excluded": {
            "$ref": "#/components/schemas/Excluded"
          },
          "facets": {
            "$ref": "#/components/schemas/Facets"
          },
          "hits": {
            "items": {
              "$ref": "#/components/schemas/Hit"
            },
            "title": "Hits",
            "type": "array"
          },
          "index_version": {
            "title": "Index Version",
            "type": "string"
          },
          "query": {
            "$ref": "#/components/schemas/QueryInfo"
          },
          "query_version": {
            "title": "Query Version",
            "type": "string"
          },
          "tokenizer_version": {
            "title": "Tokenizer Version",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "index_version",
          "tokenizer_version",
          "query_version",
          "query",
          "total",
          "excluded",
          "facets",
          "hits"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
...
      "Term": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "description": "Open set: new values may be added within /api/v1; handle a value you don't know.",
                "enum": [
                  "title",
                  "abstract"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "kind": {
            "const": "term",
            "default": "term",
            "title": "Kind",
            "type": "string"
          },
          "span": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "integer"
              },
              {
                "type": "integer"
              }
            ],
            "title": "Span",
            "type": "array"
          },
          "token": {
            "pattern": "^\\S+$",
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "span",
          "kind",
          "token",
          "field"
        ],
        "title": "Term",
        "type": "object"
      },
...
      "YearRange": {
        "properties": {
          "hi": {
            "maximum": 9999,
            "minimum": 1000,
            "title": "Hi",
            "type": "integer"
          },
          "lo": {
            "maximum": 9999,
            "minimum": 1000,
            "title": "Lo",
            "type": "integer"
          }
        },
        "required": [
          "lo",
          "hi"
        ],
        "title": "YearRange",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "openproceedings",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/parse": {
      "post": {
        "description": "02's ParseResult for `q` (debounced as the user types). It reports a parse: any well-formed body is\na 200 whose `errors` say why the query doesn't parse (`PARSE_TOO_LONG` included); only a malformed body\nis refused (422 `API_BAD_PARAM`).",
        "operationId": "parse_query",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API_METHOD_NOT_ALLOWED: the endpoint exists, but not for this method",
            "headers": {
              "Allow": {
                "description": "The methods this endpoint takes, comma-separated (RFC 9110 §10.2.1)",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API_RATE_LIMITED: this client's token bucket or its network's (IPv4 /24, IPv6 /48) can't pay for the request, a query's position-verified clauses included; or, on POST /records, the save ceiling of this client's network or of the whole instance is reached",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds until the request would be allowed",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error (spec 04 §Error handling)"
          }
        },
        "summary": "Parse Query"
      }
    },
...
    "/api/v1/search": {
      "get": {
        "description": "One page of the ranked matched set, with its total, exclusion accounting, disjunctive facets and\nhighlights. `limit` over 200 is a 422 (never clamped); a query that doesn't parse is a 422 with its\ndiagnostics.",
        "operationId": "search",
        "parameters": [
          {
            "description": "The query (spec 02 grammar). At most 2,000 Unicode code points: a longer one is 422 `PARSE_TOO_LONG`, refused before it is parsed; so is one whose canonical form (defaults written out) is longer, refused after canonicalising.",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "The query (spec 02 grammar). At most 2,000 Unicode code points: a longer one is 422 `PARSE_TOO_LONG`, refused before it is parsed; so is one whose canonical form (defaults written out) is longer, refused after canonicalising.",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "description": "`native` (this grammar) or `scholar` (Google Scholar / Publish or Perish syntax, translated).",
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "native",
              "description": "`native` (this grammar) or `scholar` (Google Scholar / Publish or Perish syntax, translated).",
              "enum": [
                "native",
                "scholar"
              ],
              "title": "Mode",
              "type": "string"
            }
          },
          {
            "description": "The order of `hits`. Never changes `total` or membership (guarantee 5).",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "relevance",
              "description": "The order of `hits`. Never changes `total` or membership (guarantee 5).",
              "enum": [
                "relevance",
                "year_desc",
                "year_asc",
                "title"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "description": "Hits to skip. Past the end is an empty page, not an error.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Hits to skip. Past the end is an empty page, not an error.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "Hits per page, 0 to 200. Over 200 is 422 `API_BAD_PARAM`, never clamped.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Hits per page, 0 to 200. Over 200 is 422 `API_BAD_PARAM`, never clamped.",
              "maximum": 200,
              "minimum": 0,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API_METHOD_NOT_ALLOWED: the endpoint exists, but not for this method",
            "headers": {
              "Allow": {
                "description": "The methods this endpoint takes, comma-separated (RFC 9110 §10.2.1)",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API_RATE_LIMITED: this client's token bucket or its network's (IPv4 /24, IPv6 /48) can't pay for the request, a query's position-verified clauses included; or, on POST /records, the save ceiling of this client's network or of the whole instance is reached",
            "headers": {
              "Retry-After": {
                "description": "Whole seconds until the request would be allowed",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API_BUSY (with Retry-After): the query needs a slow position check and every verification slot is taken; or API_INDEX_NOT_LOADED (no index loaded yet), or on POST /records API_RECORDS_STORE_FULL (neither sends Retry-After)",
            "headers": {
              "Retry-After": {
                "description": "Sent with API_BUSY: whole seconds to wait before retrying",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error (spec 04 §Error handling)"
          }
        },
        "summary": "Search"
      }
    }
  }
}